requirements.txt
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
praw==7.7.1
prawcore==2.4.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
